加一個 BaseController，把連線字串跟 DB 實體放在裡面，所有 Controller 改繼承它
回傳 JSON 字串的 Action 改成 ContentResult 回傳，並加上 "application/json"

新增 Limits (權限)

Menus 的 Level 變更型別(int -> tinyint)

把 Md5 移出，LoginController 改使用 Commons/Md5.cs

權限對應表(未完

+++ MVC/ArmyAPI/Models/Menus.cs
@@ -12,7 +12,7 @@ namespace ArmyAPI.Models
         public string Title { get; set; }
 		public int ParentIndex { get; set; }
         public string Route_Tableau { get; set; }
-		public int Level { get; set; }
+		public short Level { get; set; }
         public bool IsEnable { get; set; }
 		public bool IsFix { get; set; }
 		public DateTime AddDatetime { get; set; }
